Most of the API endpoints are implemented

+++ backend/CaffShop/Interfaces/IPurchaseService.cs
@@ -6,9 +6,9 @@ namespace CaffShop.Interfaces
 {
     public interface IPurchaseService
     {
-        Task<Purchase> PurchaseItem(int itemId, int userId);
-        Task<List<Purchase>> GetItemPurchases(int itemId);
-        Task<int> CountItemPurchases(int itemId);
-        Task<bool> IsUserPurchasedItem(int itemId, int userId);
+        Task<Purchase> PurchaseItem(long itemId, long userId);
+        Task<List<Purchase>> GetItemPurchases(long itemId);
+        Task<long> CountItemPurchases(long itemId);
+        Task<bool> IsUserPurchasedItem(long itemId, long userId);
     }
 }
